+++ model/sanity_check.go
@@ -7,7 +7,7 @@ import (
 
 type SanityCheck struct {
 	gorm.Model
-	ContentID uint       `json:"-" gorm:"uniqueIndex:cnt_blk_cid_index;index;not null"`
+	ContentID uint64     `json:"-" gorm:"uniqueIndex:cnt_blk_cid_index;index;not null"`
 	BlockCid  util.DbCID `json:"-" gorm:"uniqueIndex:cnt_blk_cid_index;index;not null"`
 	ErrMsg    string     `json:"-" gorm:"type:text;not null"`
 }
